+++ visualization/app/codeCharta/codeCharta.model.ts
@@ -18,6 +18,16 @@ export enum FileSelectionState {
 	None = "None"
 }
 
+export interface ExportCCFile {
+	projectName: string
+	apiVersion: string
+	nodes: CodeMapNode[]
+	attributeTypes: AttributeTypes | {}
+	edges: Edge[]
+	markedPackages: MarkedPackage[]
+	blacklist: BlacklistItem[]
+}
+
 export interface CCFile {
 	map: CodeMapNode
 	settings: {
@@ -64,6 +74,7 @@ export interface DynamicSettings {
 	areaMetric: string
 	heightMetric: string
 	colorMetric: string
+	distributionMetric: string
 	focusedNodePath: string
 	searchedNodePaths: Array<string>
 	searchPattern: string
@@ -112,17 +123,17 @@ export interface ColorRange {
 }
 
 export interface AttributeTypes {
-	nodes?: {
-		[key: string]: AttributeType
-	}
-	edges?: {
-		[key: string]: AttributeType
-	}
+	nodes: AttributeType[]
+	edges: AttributeType[]
 }
 
-export enum AttributeType {
-	absolute,
-	relative
+export interface AttributeType {
+	[key: string]: AttributeTypeValue
+}
+
+export enum AttributeTypeValue {
+	absolute = "absolute",
+	relative = "relative"
 }
 
 export interface Edge {