+++ visualization/app/codeCharta/util/fileDownloader.ts
@@ -1,27 +1,62 @@
 import angular from "angular"
 import * as d3 from "d3"
-import { CCFile, CodeMapNode } from "../codeCharta.model"
+import { CodeMapNode, BlacklistType, BlacklistItem, FileSettings, ExportCCFile, FileMeta, AttributeTypes } from "../codeCharta.model"
+import { DownloadCheckboxNames } from "../ui/dialog/dialog.download.component"
+import { CodeChartaService } from "../codeCharta.service"
 
 export class FileDownloader {
-	private static CC_FILE_EXTENSION = ".cc.json"
-
-	public static downloadCurrentMap(file: CCFile) {
-		const data = this.getProjectDataAsCCJsonFormat(file)
-		this.downloadData(data, this.getNewFileName(file))
+	public static downloadCurrentMap(
+		map: CodeMapNode,
+		fileMeta: FileMeta,
+		fileSettings: FileSettings,
+		downloadSettingsNames: string[],
+		fileName: string
+	) {
+		const exportCCFile: ExportCCFile = this.getProjectDataAsCCJsonFormat(map, fileMeta, fileSettings, downloadSettingsNames)
+		const newFileNameWithExtension: string = fileName + CodeChartaService.CC_FILE_EXTENSION
+		this.downloadData(exportCCFile, newFileNameWithExtension)
 	}
 
-	private static getProjectDataAsCCJsonFormat(file: CCFile) {
-		let newFileName = this.getNewFileName(file)
-
+	private static getProjectDataAsCCJsonFormat(
+		map: CodeMapNode,
+		fileMeta: FileMeta,
+		fileSettings: FileSettings,
+		downloadSettingsNames: string[]
+	): ExportCCFile {
 		return {
-			fileName: newFileName,
-			projectName: file.fileMeta.projectName,
-			apiVersion: file.fileMeta.apiVersion,
-			nodes: [this.removeJsonHashkeysAndVisibleAttribute(file.map)],
-			edges: file.settings.fileSettings.edges,
-			attributeTypes: file.settings.fileSettings.attributeTypes,
-			blacklist: file.settings.fileSettings.blacklist
+			projectName: fileMeta.projectName,
+			apiVersion: fileMeta.apiVersion,
+			nodes: [this.removeJsonHashkeysAndVisibleAttribute(map)],
+			attributeTypes: this.getAttributeTypesForJSON(fileSettings.attributeTypes),
+			edges: downloadSettingsNames.includes(DownloadCheckboxNames.edges) ? fileSettings.edges : [],
+			markedPackages: downloadSettingsNames.includes(DownloadCheckboxNames.markedPackages) ? fileSettings.markedPackages : [],
+			blacklist: this.getBlacklistToDownload(downloadSettingsNames, fileSettings.blacklist)
 		}
+	}
+
+	private static getBlacklistToDownload(downloadSettingsNames: string[], blacklist: BlacklistItem[]): BlacklistItem[] {
+		let mergedBlacklist: BlacklistItem[] = []
+
+		if (downloadSettingsNames.includes(DownloadCheckboxNames.hides)) {
+			mergedBlacklist.push(...this.getFilteredBlacklist(blacklist, BlacklistType.hide))
+		}
+
+		if (downloadSettingsNames.includes(DownloadCheckboxNames.excludes)) {
+			mergedBlacklist.push(...this.getFilteredBlacklist(blacklist, BlacklistType.exclude))
+		}
+		return mergedBlacklist
+	}
+
+	private static getAttributeTypesForJSON(attributeTypes: AttributeTypes): AttributeTypes | {} {
+		if (attributeTypes.edges.length === 0 && attributeTypes.nodes.length === 0) {
+			return {}
+		} else {
+			return attributeTypes
+		}
+	}
+
+	private static getFilteredBlacklist(blacklist: BlacklistItem[], type: BlacklistType): BlacklistItem[] {
+		return blacklist.filter(x => x.type == type)
 	}
 
 	private static removeJsonHashkeysAndVisibleAttribute(map: CodeMapNode) {
@@ -32,32 +67,7 @@ export class FileDownloader {
 		return copy
 	}
 
-	private static getNewFileName(file: CCFile): string {
-		return this.getFileNameWithoutTimestamp(file.fileMeta.fileName) + this.getNewTimestamp() + FileDownloader.CC_FILE_EXTENSION
-	}
-
-	private static getNewTimestamp(): string {
-		const date: Date = new Date()
-		return (
-			"_" + date.getFullYear() + "-" + (date.getMonth() + 1) + "-" + date.getDate() + "_" + date.getHours() + "-" + date.getMinutes()
-		)
-	}
-
-	private static getFileNameWithoutTimestamp(fileName: string): string {
-		const dateRegex: RegExp = /\_\d{4}\-\d{1,2}\-\d{1,2}\_\d{1,2}\-\d{1,2}\./
-
-		if (dateRegex.test(fileName)) {
-			return fileName.substring(0, dateRegex.exec(fileName).index)
-		} else if (fileName.includes(FileDownloader.CC_FILE_EXTENSION)) {
-			return fileName.substring(0, fileName.search(FileDownloader.CC_FILE_EXTENSION))
-		} else if (fileName.includes(".json")) {
-			return fileName.substring(0, fileName.search(".json"))
-		} else {
-			return fileName
-		}
-	}
-
-	private static downloadData(data, fileName) {
+	private static downloadData(data, fileName: string) {
 		let dataJson = data
 		if (typeof data === "object") {
 			dataJson = angular.toJson(data, 4)