+++ CHANGELOG.md
@@ -6,6 +6,35 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ## [unreleased]
+### Added
+
+### Changed
+
+### Removed
+
+### Fixed
+
+### Chore
+
+## [1.26.0] - 2019-06-14
+
+### Added
+
+- FileExtensionBar to show file-distribution of chosen metric #495
+- Added Pop-up dialog before downloading file to set filename and see what data will be stored #523
+- Sum icon is now displayed on the left of the metric value #364
+
+### Changed
+
+### Removed
+
+### Fixed
+
+- Fix set default ColorRange when resetting color section #560
+
+### Chore
+
+## [1.25.1] - 2019-05-30
 
 ### Added